authentication_microservice/pyproject.toml
[project]
name = "authentication-microservice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.136.1",
    "pwdlib[argon2]>=0.3.0",
    "pydantic[email]>=2.13.3",
    "pyjwt>=2.12.1",
]
